python/0215 - Recipe Builder API/requirements.txt
https://github.com/hyperskill/hs-test-python/archive/release.tar.gz
Flask-SQLAlchemy
marshmallow
